Refactored notification icons and prevented a crash condition
